requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
dj-rest-auth==7.0.1
django==4.2.21
django-allauth==65.8.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
oauthlib==3.2.2
pycparser==2.22
pyjwt==2.9.0
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.25.1
sqlparse==0.5.3
typing-extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
